chore: cleanup context set and get BREAKING CHANGE: removed setTime on context
