--- src/lib/components/key.ts
@@ -1,1 +1,0 @@
-export const key = Symbol();
